requirements.txt
requests
python-dotenv
Flask

